requirements.txt
sqlalchemy==2.0.49
alembic==1.18.4
psycopg2-binary==2.9.11
pydantic==2.13.3
pydantic-settings==2.13.1
python-dotenv==1.2.2
requests==2.33.1
httpx==0.28.1
beautifulsoup4==4.14.3
feedparser==6.0.12
lxml==6.1.0
python-dateutil==2.9.0.post0
fastapi==0.135.3
uvicorn[standard]==0.43.0
python-telegram-bot==22.5
google-genai==1.73.1
pytest==9.0.3
